+++ spec/kf5/configuration_spec.rb
@@ -1,0 +1,8 @@
+RSpec.describe KF5::Configuration do
+  subject { properties }
+  let(:properties) { KF5::Configuration.new(:domain => 'test') }
+  
+  describe "#initialize" do
+    its(:domain) { should eq 'test' }
+  end
+end