+++ spec/kf5/view_helper_spec.rb
@@ -1,0 +1,9 @@
+RSpec.describe KF5::ViewHelper do
+  include KF5::ViewHelper
+  include ActionView::Helpers::AssetTagHelper
+
+  describe "#kf5_tag" do
+    subject { kf5_tag }
+    it { should eq "<script src=\"/javascripts/kf5.js\" id=\"kf5-provide-supportBox\" kf5-domain=\"domain.kf5.com\"></script>" }
+  end
+end