+++ spec/spec_helper.rb
@@ -1,15 +1,19 @@
 require 'simplecov'
 require 'coveralls'
 
+SimpleCov.add_filter "/spec/"
+SimpleCov.add_filter "engine.rb"
+
 if ENV["COVERAGE"]
-	SimpleCov.start 
-else
-	SimpleCov.formatter = Coveralls::SimpleCov::Formatter
-	Coveralls.wear!
+  SimpleCov.start
+elsif ENV["COVERALLS"]
+  SimpleCov.formatter = Coveralls::SimpleCov::Formatter
+  Coveralls.wear!
 end
 
 $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 require 'kf5'
+require 'action_view'
 require 'rspec/its'
 require 'ostruct'