+++ spec/support/controller.rb
@@ -1,7 +1,7 @@
 class Controller
   attr_accessor :redirect_url
 
-  include KF5::Helper
+  include KF5::ControllerHelper
 
   def redirect_to(url)
     @redirect_url = url